# faster `methods()`