This is faster, much more correct, and somewhat more general than https://github.com/JuliaLang/julia/pull/38061, assuming this is acceptable. Not quite as fast as v1.5, but enough closer hopefully.

```
julia> @time ms = collect(methods(mul!));
  0.001091 seconds (1.39 k allocations: 101.562 KiB) # v1.5.2
  0.069216 seconds (81.95 k allocations: 6.425 MiB, 9.15% gc time) # this PR
  0.08 seconds # other PR (approximate)
  3.400419 seconds (7.26 M allocations: 718.603 MiB, 1.80% gc time) # master
```

(notes: the first commit is best viewed with ignore white space. the second commit is putting a piece of 901f1e283e75a17ae669d88500307bd7cc8c1758 behind a condition.)